[repositories] bookworm: expand the glibc workaround to every architecture
Unfortunately the faccessat2() issue with the outdated seccomp we have in
our build infra's hosts (outside our control) means that the stock glibc 2.33
packaged don't work.

This has been resolved in our variant (which will be used only in these docker
images, *not* on devices!) which are now compiled for every supported
architecture.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>